refactor GlobalExceptionHandler
