--- vcf-dumper/vcf-dumper-ws/src/main/java/uk/ac/ebi/eva/vcfdump/server/ApiError.java
@@ -1,35 +1,0 @@
-/*
- * Copyright 2018 EMBL - European Bioinformatics Institute
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package uk.ac.ebi.eva.vcfdump.server;
-
-class ApiError {
-
-    private String message;
-    private String cause;
-
-    ApiError(String message, String cause) {
-        this.message = message;
-        this.cause = cause;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public String getCause() {
-        return cause;
-    }
-}
